added SSH key configuration and sftp-sender files
ssh_setup.py - generates an SSH key pair for the current user, saves it in the .ssh directory. It then updates the SSH config file with an entry for a specified remote host, linking the new key for authentication.

sftp_uploader.py - uploads files to an SFTP server using SSH key authentication. It dynamically retrieves the current user's username, checks if the required private key exists, and connects to the server using the paramiko library. It creates the necessary remote directories and uploads the specified files to the server. If any file does not exist, it is skipped. pip install paramiko is needed.

run_sftp_uploader - gathers files from a specified directory on the user's Desktop/logs(for testing purposes) and uploads them to an SFTP server using the `upload_files` function from the sftp_uploader.py script. It checks if the directory exists and retrieves all files for upload. If no files are found, it exits. The script will later likely be transformed into the main script for PSWatchdog.

The configurations for the remote server and SSH settings will be updated next time.